Fix parsing of mode lines with trailing space
If a patch is passed through a system that converts line endings to
'\r\n', mode lines end up with trailling whitespace that confuses
ParseInt. In Git, this is avoided by using strtoul() for parsing, which
stops at the first non-digit character.

Changing line endings in patch content itself will cause other problems
so it is best to avoid this transform, but if it does happen, it
shouldn't cause a parse error.